Migrate server and test files to TypeScript
- Updated package.json to include TypeScript and related devDependencies.
- Archived popup.js, server.js, and test.js, migrating their logic to TypeScript files (popup.ts, server.ts, test.ts).
- Created tsconfig.json for TypeScript configuration.
- Added Vite configuration for building the server and test files.
