+++ extension/index.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="UTF-8" />
+    <title>MeterX Extension Build</title>
+  </head>
+  <body>
+    <script type="module" src="popup.ts"></script>
+  </body>
+</html>
