+++ package.json
@@ -149,13 +149,16 @@
     "@eslint/css": "^0.7.0",
     "@eslint/js": "^9.26.0",
     "@eslint/markdown": "^6.4.0",
+    "@types/node": "^24.0.13",
     "eslint": "^9.26.0",
     "globals": "^16.1.0",
-    "typescript-eslint": "^8.32.0"
+    "typescript": "^5.8.3",
+    "typescript-eslint": "^8.32.0",
+    "vite": "^7.0.4"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "lint:fix":"eslint . --ext .js"
+    "lint:fix": "eslint . --ext .js"
   },
   "keywords": [],
   "author": "",
