+++ server/vite.config.ts
@@ -1,0 +1,15 @@
+import { defineConfig } from 'vite';
+
+export default defineConfig({
+  root: '.',
+  build: {
+    outDir: 'dist',
+    emptyOutDir: true,
+    rollupOptions: {
+      input: {
+        server: 'server.js',
+        test: 'test.js'
+      }
+    }
+  },
+});
